#97 DocID in local storage for persistency
